requirements.txt
ccxt
pandas
numpy
requests
python-dotenv
ta
